Add files via upload
Minor changes. Enable not necessarily needed and is commented out in the version I am using. This may be dependent on the environment and how permissions are configured on the switches. You may need to adjust as needed until/unless code is written to make input of enable password optional